Commit 2025-07-15

+++ src/routes/(protected)/api/admin/calendar/[id]/+server.ts
@@ -1,25 +1,27 @@
-import {prisma, prismaErrorHandler} from '@lib/utils'
+import { pecahArray, prisma, prismaErrorHandler } from '@lib/utils'
 import { error, json } from '@sveltejs/kit'
 
-export async function GET({params}){
+export async function GET({ params }) {
     const id = params.id
     const req = await prisma.calendar.findUnique({
-        where:{
+        where: {
             calendar_id: id
         }
     })
     return json(req)
 }
 
-export async function DELETE({params}){
+export async function DELETE({ params, locals }) {
     try {
         const id = params.id
+        const { userProfile } = locals
+        if (!pecahArray(userProfile.access_calendar, "D")) throw new Error("Cant delete Calendar, because you have no authorization")
         const req = await prisma.calendar.delete({
-            where:{
+            where: {
                 calendar_id: id
             }
         })
-        return json({message:"Data successfully deleted"})
+        return json({ message: "Data successfully deleted" })
     } catch (err) {
         error(500, prismaErrorHandler(err))
     }
